requirements.txt
pyqt5
pyqt5-tools
pymongo
python-dotenv
bcrypt